requirements.txt
streamlit 
python-dotenv
openai
langchain_community
langchain_openai
pypdf
faiss-cpu
pillow